requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
open_clip_torch
torch
torchvision
pandas
pillow
python-multipart
pydantic
datasets
tqdm
matplotlib
python-dotenv
google-generativeai
aiofiles
